requirements.txt
gradio 
sentence_transformers 
faiss-cpu 
ffmpeg-python 
pandas 
pyarrow 
google-generativeai
openai